+++ packages/stripe_js/lib/src/api/payment_methods/payment_methods.dart
@@ -1,4 +1,5 @@
 export 'create_payment_method_data.dart';
+export 'create_payment_method_with_elements_data.dart';
 export 'payment_method.dart';
 export 'payment_method_details.dart';
 export 'payment_method_params.dart';