+++ packages/stripe_js/lib/src/js/elements/elements_submit.dart
@@ -1,0 +1,11 @@
+import 'dart:js_interop';
+
+import 'package:stripe_js/stripe_api.dart';
+import 'package:stripe_js/stripe_js.dart';
+
+extension ExtensionElementsSubmit on StripeElements {
+  Future<void> submit() => _submit().toDart;
+
+  @JS('submit')
+  external JSPromise<JSAny?> _submit();
+}