+++ packages/stripe_platform_interface/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 11.0.1
+
+**Features**
+- add methods for deferred payment on web
+
 ## 11.0.0
 **Breaking changes**
 - Add support for intentmode setupMode. If you used `IntentMode before` change it to `IntentMode.paymentMode` instead.